attieke_ivoir_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
Pillow>=10.2.0
gunicorn>=21.2.0
django-filter>=23.5
drf-spectacular>=0.27.1
